requirements.txt
# Mask2Derm — Python dependencies
# Tested with Python 3.10+

# --- Core deep learning ---
torch>=2.1.0
torchvision>=0.16.0

# --- Diffusion / ControlNet ---
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0

# --- Dataset management ---
datasets>=2.18.0
huggingface_hub>=0.21.0

# --- Image processing ---
opencv-python>=4.9.0
Pillow>=10.0.0

# --- Data & utilities ---
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
omegaconf>=2.3.0
tqdm>=4.66.0

# --- Evaluation metrics ---
clean-fid>=0.1.35
torch-fidelity>=0.3.0

# --- Visualization & logging ---
matplotlib>=3.7.0
wandb>=0.16.0

# --- Optional: Kaggle dataset download ---
# kaggle>=1.6.0
